servers/azure-ai-foundry-agents-mcp/requirements.txt
mcp>=1.3.0
azure-identity>=1.17.0
azure-ai-projects>=1.0.0
